requirements.txt
annoy==1.17.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
executing==0.8.3
faiss-cpu==1.7.2
fastapi==0.75.2
h11==0.13.0
idna==3.3
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
jedi==0.18.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.6
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.22.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pyrsistent==0.18.1
python-dotenv==0.20.0
pyzmq==22.3.0
Send2Trash==1.8.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
terminado==0.13.2
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.2.0
uvicorn==0.17.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
